requirements - Copy.txt
flask>=2.0.1
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
scikit-image>=0.21.0
scikit-learn>=1.0.2
werkzeug>=2.0.1
